fix: ignore unexpected DNS response types (#3)
DNS server responses can mismatch the expected query type.
For example, a DNS server can respond with an A record
even if TXT was requested.

For our purposes, we generally want to ignore these responses.
This is similar to two bugs the MongoDB C driver experienced
in the past. (The code in this package was adapted from the
server code, rather than the C driver code. I think the
server code does not typically encounter these situations
and may also be just fine when they happen.)

Refs: https://jira.mongodb.org/browse/CDRIVER-2789
Refs: https://jira.mongodb.org/browse/CDRIVER-3628